tests/wpt/tests/css/css-borders/corner-shape/corner-invalid.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Borders and Box Decorations 4 Test: Parsing 'corner' with invalid values</title>
<link rel="author" title="Sebastian Zartner" href="mailto:sebastianzartner@gmail.com">
<link rel="help" href="https://drafts.csswg.org/css-borders-4/#corner-shaping">
<meta name="assert" content="This test checks that 'corner' supports only a combination of 'border-radius' and 'corner-shape' values.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>

<script>
test_invalid_value("corner", "auto");
test_invalid_value("corner", "none");
test_invalid_value("corner", "scoop");
test_invalid_value("corner", "3% normal");
test_invalid_value("corner", "round round round round round");
test_invalid_value("corner", "-1px");
test_invalid_value("corner", "10px");
test_invalid_value("corner", "4px 12px");
test_invalid_value("corner", "4px 12px / normal");
test_invalid_value("corner", "4px 2% 1em 4% 2px");
test_invalid_value("corner", "4px / 2px 4% 1em 2% 2px");
test_invalid_value("corner", "4px / angle");
test_invalid_value("corner", "angle / 4px");
test_invalid_value("corner", "4px angle 2% round");
</script>
